% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_funs.R
\name{fuzzy_substring_distance}
\alias{fuzzy_substring_distance}
\title{Fuzzy substring distance}
\usage{
fuzzy_substring_distance(texts, pattern, ...)
}
\arguments{
\item{texts}{vector of texts}

\item{pattern}{pattern to look for}

\item{...}{arguments passed on to \code{stringdist::stringdist()}}
}
\value{
distance per text
}
\description{
Fuzzy substring distance
}
