% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdu_funs_generic.R
\name{plot.rdu}
\alias{plot.rdu}
\title{Plot an rdu object}
\usage{
\method{plot}{rdu}(
  x,
  stage = 0,
  file = NULL,
  image = FALSE,
  ocr_alpha = 1,
  width_ext = 1.5,
  ...
)
}
\arguments{
\item{x}{receipt object}

\item{stage}{stage of the processing. By default 0, meaning the latest.}

\item{file}{output file}

\item{image}{should the image be plotted?}

\item{ocr_alpha}{alpha transparency for ocr text when plotted on top of the image}

\item{width_ext}{extention of the width}

\item{...}{not used}
}
\description{
Plot an rdu object
}
