% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdu_funs_generic.R
\name{print.rdu}
\alias{print.rdu}
\title{Print an rdu object}
\usage{
\method{print}{rdu}(x, ...)
}
\arguments{
\item{x}{\code{rdu} object}

\item{...}{not used}
}
\description{
Print an rdu object
}
