% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdu_funs_init_check.R
\name{rdu_check_image}
\alias{rdu_check_image}
\title{Check the receipt image}
\usage{
rdu_check_image(r, include_image = FALSE)
}
\arguments{
\item{r}{\code{rdu} object}

\item{include_image}{should the image itself be included in the returned \code{rdu} object? \code{FALSE} by default}
}
\value{
an rdu object. In case of an error, the attribute \code{"abort"} is added, which contains the error message
}
\description{
Check the receipt image
}
