% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdu_funs_init_check.R
\name{rdu_check_OCR}
\alias{rdu_check_OCR}
\title{Check the OCR file}
\usage{
rdu_check_OCR(r)
}
\arguments{
\item{r}{\code{rdu} object}
}
\value{
an rdu object. In case of an error, the attribute \code{"abort"} is added, which contains the error message
}
\description{
Check the OCR file. Currently checks if the variables text, left, top, width, block_num, par_num and line_num are present. The latter three are probably typical for Tesseract. This function needs to be updated to support Paddle output.
}
