% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdu_funs_process.R
\name{rdu_check_table}
\alias{rdu_check_table}
\title{Check table}
\usage{
rdu_check_table(r)
}
\arguments{
\item{r}{\code{rdu} object}
}
\value{
an rdu object. In case of an error, the attribute \code{"abort"} is added, which contains the error message
}
\description{
Check table
}
