% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdu_funs_process.R
\name{rdu_compute_line_numbers}
\alias{rdu_compute_line_numbers}
\title{Compute line numbers}
\usage{
rdu_compute_line_numbers(r)
}
\arguments{
\item{r}{\code{rdu} object}
}
\value{
an rdu object. In case of an error, the attribute \code{"abort"} is added, which contains the error message
}
\description{
Compute line numbers
}
