% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdu_funs_process.R
\name{rdu_create_lines}
\alias{rdu_create_lines}
\title{Create lines}
\usage{
rdu_create_lines(r)
}
\arguments{
\item{r}{\code{rdu} object}
}
\value{
an rdu object. In case of an error, the attribute \code{"abort"} is added, which contains the error message
}
\description{
Create lines
}
