% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdu_funs_process.R
\name{rdu_detect_shop}
\alias{rdu_detect_shop}
\title{Detect shop}
\usage{
rdu_detect_shop(r, templates, params)
}
\arguments{
\item{r}{\code{rdu} object}

\item{templates}{the shop templates}

\item{params}{the parameters object}
}
\value{
an rdu object. In case of an error, the attribute \code{"abort"} is added, which contains the error message
}
\description{
Detect shop
}
