% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdu_funs_process.R
\name{rdu_get_product_lines}
\alias{rdu_get_product_lines}
\title{Get product lines}
\usage{
rdu_get_product_lines(r)
}
\arguments{
\item{r}{\code{rdu} object}
}
\value{
an rdu object. In case of an error, the attribute \code{"abort"} is added, which contains the error message
}
\description{
Get product lines
}
