% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdu_funs_init_check.R
\name{rdu_init}
\alias{rdu_init}
\title{Initiate a rdu object}
\usage{
rdu_init(file_image, file_OCR, meta = list(), shop = NA, total_discount = NA)
}
\arguments{
\item{file_image}{path the the image file (required)}

\item{file_OCR}{path the the OCR file (required)}

\item{meta}{meta information, not used yet}

\item{shop}{Set the shop template (use \code{NA} to determine automatically later on)}

\item{total_discount}{Total price paid, so with discounts already applied. If not specified it is read from the receipt.}
}
\value{
an rdu object
}
\description{
Initiate a rdu object
}
