% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdu_funs_process.R
\name{rdu_process_table}
\alias{rdu_process_table}
\title{Process table}
\usage{
rdu_process_table(r)
}
\arguments{
\item{r}{\code{rdu} object}
}
\value{
an rdu object. In case of an error, the attribute \code{"abort"} is added, which contains the error message
}
\description{
Process table
}
