% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdu_io.R
\name{rdu_read}
\alias{rdu_read}
\title{Read rdu object from json file}
\usage{
rdu_read(file)
}
\arguments{
\item{file}{json file}
}
\value{
\code{rdu} object
}
\description{
Read rdu object from json file
}
