% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdu_io.R
\name{rdu_write}
\alias{rdu_write}
\title{Write rdu object to json file}
\usage{
rdu_write(x, file, include_throughput = FALSE)
}
\arguments{
\item{x}{\code{rdu} object}

\item{file}{target json file name}

\item{include_throughput}{should intermediate output be included?}
}
\value{
\code{x}
}
\description{
Write rdu object to json file
}
